Introduce WriteOrPanic trait for write!-ing without unwrap. (#771)
This commit introduces a new utility trait WriteOrPanic which provides the
method write_fmt that is used by write! macro and friends. But because it
doesn’t return a result, it allows using the macro without unwrap or
question mark.

The trait is implemented for String and Vec<u8> and used where needed.

Much like the trait’s name suggests, these implementations panic if
formatting fails. That’s fine, as we only use with data that can be
formatted and with types where writing itself can never fail (which I think
is the actual reason for returning a result). Also, we’ve been unwrapping
for quite some time so, clearly, this is working.