Do not exit on malformed RRDP Base 64 data. (#781)
This commit fixes an oversight in error handling in the RRDP collector that
causes Routinator to exit if it encountered malformed Base 64 in RRDP
snapshot and delta files.

It does this by re-appropriating the existing exception in MaxSizeRead to
not exit when data was too large and extending it to all other read errors
happening. This is fine because the encoded data is currently collected into
a vec before any of this happens, so reading cannot fail other than from
malformed Base 64 data.

This commit only does the minimal changes to be applied to the 0.11 series.
Before applying to the main branch, MaxSizeRead should probably be given a
better name.

This commit can be tested by using the TAL currently available at
https://routinator.do.nlnetlabs.nl/test/rrdp/ta.tal.

This issue fixed by this commit has been assigned CVE-2022-3029.