Do not exit on malformed RRDP Base 64 data. (#784)
This commit fixes an oversight in error handling in the RRDP collector that
cause Routinator to exit if it encountered malformed Base 64 in RRDP
snapshot and delta files.

It does this by re-appropriating the existing exception in MaxSizeRead to
not exit when data was too large and extending it to all other read errors
happening. This is fine because the encoded data is currently collected into
a vec before any of this happens, so reading cannot fail other than from
malformed Base 64 data.

This PR can be tested by using the TAL currently available at
https://routinator.do.nlnetlabs.nl/test/rrdp/ta.tal.

This is the same PR as #781 accept applied to the main branch. It fixes
CVE-2022-3029.