Add an rrdp-fallback option to configure fallback to rsync. (#799)
This commit introduces a new command line and config option "rrdp-fallback"
that decides when to fall back to rsync for a failed RRDP update. The
available policies are "never" for never, "stale" for fallback after a
certain time of falling the RRDP update, and "new" for fallback when RRDP has
never succeeded. The default is "stale" which reflects the current behavior.

Co-authored-by: Alex Band <alex@nlnetlabs.nl>
Co-authored-by: Luuk Hendriks <mail@luukhendriks.eu>