Refer to (standalone) EE certs as router certs. (#854)
This PR changes renames functions and attributes etc. that refer to
standalone end-entity certificates to refer to them as ‘router certs’ to
better distinguish them from EE certificates that are part of signed
objects.

The only external change is in the HTTP status API where the `validEECerts`
member has been renamed to `validRouterCerts` but the old name has been
retained for now.

This is a breaking change.