Disable more ASPA code without the aspa feature. (#878)
This PR completely disables processing of .asa files and removes output of
ASPA related metrics if the aspa feature is not enabled.

Previously, the change in ASPA format and the presence of a number of
old-format ASPA objects in the repositories lead to a sequence of confusing
error messages. Now .asa files are simply ignored when encountered.