Don’t have the RTR listener fail if a socket fails after accept. (#939)
This PR changes the RTR listener to not fail the accept loop when setting up
a stream after accepting it fails. Instead it will quietly drop the stream
in this case and keep going.

As part of this, the PR also drops tokio-stream as a dependency and
implements its own listener stream.

This PR fixes CVE-2024-1622 reported by Yohei Nishimura, Atsushi Enomoto,
Ruka Miyachi; Internet Multifeed Co., Japan.