Store manifest number and check for regressions. (#946)
This PR adds a check for manifest number regressions when validating a
collected publication point. It stores the manifest number for each manifest
and checks against it when collecting a new manifest. If the latter’s number
has not increased, it falls back to the stored manifest. This behaviour is
mandated by RFC 9286.

The PR changes the data stored for manifests and thus updates the
StoredManifest version to 1. In order to avoid an endless stream of error
messages after an upgrade, it downgrades the logged message when
encountering an malformed StoredManifest to DEBUG.