Limit the maximum number of deltas in an RRDP notification. (#961)
This PR limits the number of entries in the notification file of an RRDP
repository that Routinator is willing to consider. This can be set through
the new rrdp-max-delta-list-len configuration value which defaults to a
hopefully safe 500.

If this number is exceeded, the delta list will be considered to be empty
and ignore, resulting in snapshots to be used.