Improve archive file sizes. (#982)
This PR improves the growth of RRDP archive files. It tries to re-use
existing empty space when updating objects. It also pads all objects to a
multiple of 256 bytes to make it more likely that empty space can be reused.

The PR also adds a new command archive-stats that prints statistics for an
archive file.

Finally, the PR raises the Minimum Supported Rust Version to 1.73 to be able
to use u64::next_multiple_of.