Restructure limiting of ASPA provider set sizes. (#996)
This PR changes how the ASPA provider ASN set is limited. It drops the
aspa-provider-limit config option again and instead hard-codes the limit
to 16380 as per the latest version of the RTR draft. Since this limit can
be exceeded when merging multiple ASPAs, the check has been moved to the
very end when the new snapshot is created which also avoid having to track
customer ASNs. A new metric was added to count the ASPA customer ASNs
ignored this way.