Add a special append archive for RRDP snapshot updatess. (#1035)
This PR adds a flavour of the archive tailored to only appending objects to
an empty archive. This allows significantly speeding up RRDP snapshot
updates by skipping checks for existing objects and keeping the index on
memory instead of writing to it on each append.

Anecdotally, this change speeds up a snapshot update of the RIPE NCC
repository from taking 35 seconds to 15 seconds (basically the time
downloading the file takes) on my laptop.