Remove rrdp-keep-responses feature. (#1055)
This PR removes the feature to store RRDP responses via the
rrdp-keep-responses config option.

The feature is very rarely used but is somewhat tricky to get right and
convenient. Instead, an HTTP proxy such as mitmproxy can be used to capture
RRDP exchanges.

This is most definitely a breaking change.