Log repository issues separately. (#1054)
This PR introduces separate logs for each repository and publication point
and puts log messages for them there. By default, these messages are not
sent to the process log any more. This is in response to many people getting
worried about these issues that are not under their control and can be
ignored.

The messages are available both via the /status and /api/v1/status HTTP
endpoints grouped by the repository or publication point which makes it much
easier to see what has issues.

The new log-repository-issues configuration option allows sending the issues
to the process log as well, restoring the previous behaviour for people who
prefer it.