Add rrdp-read-timeout and revise timeout defaults. (#1061)
This PR adds a new rrdp-read-timeout config parameter which sets the
timeout for all network operations related to RRDP. The default for this
timeout is 10 seconds. As this is also used for the initial connect, this
will make broken servers timeout much more quickly.

In addition, the PR changes the default for the rrdp-timeout value – which
concerns the full exchange – to 600 seconds as we have seen some
repositories time out the snapshot with the previous default of 300
seconds. We think it beneficiary to give them more time to finish the
snapshot update rather than falling back to rsync and then trying again
and failing again after five minutes for every validation run.