Drop corrupt stored manifests. (#1059)
This PR changes the mechanism to check whether a newly collected manifest
is newer than a stored one. If it is determined that the stored manifest
is newer, we check whether the stored data has been corrupted in the
relevant parts and if so delete it, accepting the collected manifest
anyway.

This PR is a fix for an issue where an accidentally or maliciously
manipulated locally stored manifest could block update of a legitimate new
manifest. This issue was reported by Zizhi Shang, Zhechao Lin, Jiahao Cao,
Yangyang Wang, Mingwei Xu of the Institute for Network Sciences and
Cyberspace (INSC), Tsinghua University.